router: move case classes to object
