+++ eclair-core/src/main/scala/fr/acinq/eclair/Eclair.scala
@@ -34,7 +34,8 @@ import fr.acinq.eclair.payment._
 import fr.acinq.eclair.payment.receive.MultiPartHandler.ReceivePayment
 import fr.acinq.eclair.payment.relay.Relayer.{GetOutgoingChannels, OutgoingChannels, UsableBalance}
 import fr.acinq.eclair.payment.send.PaymentInitiator.{SendPaymentRequest, SendPaymentToRouteRequest, SendPaymentToRouteResponse}
-import fr.acinq.eclair.router._
+import fr.acinq.eclair.router.Router._
+import fr.acinq.eclair.router.{NetworkStats, RouteCalculationHandlers}
 import fr.acinq.eclair.wire.{ChannelAnnouncement, ChannelUpdate, NodeAddress, NodeAnnouncement}
 import scodec.bits.ByteVector
