+++ eclair-core/src/main/scala/fr/acinq/eclair/payment/PaymentPacket.scala
@@ -21,7 +21,7 @@ import fr.acinq.bitcoin.ByteVector32
 import fr.acinq.bitcoin.Crypto.{PrivateKey, PublicKey}
 import fr.acinq.eclair.channel.{CMD_ADD_HTLC, Upstream}
 import fr.acinq.eclair.crypto.Sphinx
-import fr.acinq.eclair.router.{ChannelHop, Hop, NodeHop}
+import fr.acinq.eclair.router.Router.{ChannelHop, Hop, NodeHop}
 import fr.acinq.eclair.wire._
 import fr.acinq.eclair.{CltvExpiry, CltvExpiryDelta, Features, MilliSatoshi, UInt64, randomKey}
 import scodec.bits.ByteVector
