+++ eclair-core/src/main/scala/fr/acinq/eclair/router/NetworkStats.scala
@@ -18,6 +18,7 @@ package fr.acinq.eclair.router
 
 import com.google.common.math.Quantiles.percentiles
 import fr.acinq.bitcoin.Satoshi
+import fr.acinq.eclair.router.Router.PublicChannel
 import fr.acinq.eclair.wire.ChannelUpdate
 import fr.acinq.eclair.{CltvExpiryDelta, MilliSatoshi}
