+++ eclair-core/src/test/scala/fr/acinq/eclair/router/RoutingSyncSpec.scala
@@ -27,6 +27,7 @@ import fr.acinq.eclair.crypto.TransportHandler
 import fr.acinq.eclair.io.Peer.PeerRoutingMessage
 import fr.acinq.eclair.router.Announcements.{makeChannelUpdate, makeNodeAnnouncement}
 import fr.acinq.eclair.router.BaseRouterSpec.channelAnnouncement
+import fr.acinq.eclair.router.Router.{Data, GossipDecision, PublicChannel, SendChannelQuery, State}
 import fr.acinq.eclair.router.SyncHandlers._
 import fr.acinq.eclair.transactions.Scripts
 import fr.acinq.eclair.wire._
